implement rudimentary matrix rtc support with the m.call application type
This introduces a new event handler which listens for matrix rtc session memberships in any room the bot is joined to.
Afterwards, it joins the call, waits 10 seconds, then leaves.

The following issues still have to be fixed, before this stage is completed and I can start working on how to stream media files:

* find our livekit configuration from the server
* create our own call application type, because right now we copy the values from the person who joined the session
  * in particular, configure our call id, matrix rtc livekit foci list and prefered focus
* and maybe not listen to this event via the before mentioned handler at all,
just join a call if the bot recieves the "$BOT_PREFIX stream" command, where $BOT_PREFIX is taken from
the configuration file, because after all, maybe people don't want the bot to join every time they start the call.
Another alternative here is to just allow either behaviour via the configuration if that's something people want