remove the shortcuts taken in the beginning to join a matrix rtc session
The bot now retrieves the rtc foci list from .well-known as per the spec, composing it with the last state event we got under that key.
So, the bot now can propose the livekit instance of its homeserver instead of copying the one sent by the initiating party.

Bugs: currently, the livekit alias is sent as the empty string, because I'm not yet sure what that is and where to get it from. Perhaps from the /sfu/get endpoint, but that remains to be seen