add a type to semantically represent the difference between livekit member IDs and their matrix components
This is a small change, but it essentially allows the to_device event handler to map sent keys to the user_id, device_id touple, in livekit membership space.
This is now only used for showing information related to the user who sent the event, how their livekit identity would look like and also their key. However, in the future it could be used for stronger guarantees, like having it in a per-room map of all participants and ensuring that this information comes from the device event handler and nowhere else